refactor architecture of index export in date picker library. Separate container logic from component display file
