+++ react/lib/datepicker/index.ts
@@ -1,0 +1,5 @@
+import DatePickerContainer from "./DatePickerContainer";
+
+export { DatePickerContainer as DatePicker };
+
+
